pyproject.toml
[tool.poetry]
name = "chat2desk-test"
version = "0.1.0"
description = "Тестовое задание для chat2desk"
authors = ["Павел Костюк <pavkos0725@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
